refactor: Util패키지 생성하여 RacingCarUtil클래스 이동하였음
