+++ src/main/java/racingcar/model/RacingCars.java
@@ -3,7 +3,7 @@ package racingcar.model;
 import java.util.ArrayList;
 import java.util.List;
 
-import racingcar.RacingCarUtil;
+import racingcar.util.RacingCarUtil;
 
 public class RacingCars {
 	List<RacingCar> racingCars = new ArrayList<>();
